+++ hpc/nvidia/cuda/samples/cudaDeviceGetP2PAttribute.cu
@@ -1,0 +1,1 @@
+../../../metax/samples/hcDeviceGetP2PAttribute.hpcc