+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to
 #### Added
 - ascend: add abbreviation.md;
 - metax: add hcGetDeviceProperties.hpcc, cuda2hccl.h, hcDeviceGetP2PAttribute.hpcc, symlink dim3.hpcc;
-- cuda: add dim3.cu;
+- cuda: add dim3.cu, symlink cudaDeviceGetP2PAttribute.cu;
 - nvidia: add README.Nsight.md from ostools;
 #### Changed
 #### Deprecated